split tests similarly to cni ones
